+++ packages/headless/src/controllers/product-listing/facet/headless-product-listing-facet.ts
@@ -9,6 +9,8 @@ import {
   FacetSearchOptions,
   FacetSearchState,
   FacetState,
+  CoreFacetState,
+  CoreFacet,
   FacetValue,
   SpecificFacetSearchResult,
 } from '../../core/facets/facet/headless-core-facet';
@@ -41,6 +43,8 @@ export {
   FacetProps,
   Facet,
   FacetState,
+  CoreFacetState,
+  CoreFacet,
   FacetSearch,
   FacetSearchState,
   SpecificFacetSearchResult,